Use a dropdown instead of a modal
