+++ packages/edit-site/src/components/navigation-sidebar/navigation-panel/style.scss
@@ -93,10 +93,3 @@
 	padding-top: $grid-unit-05;
 	color: $gray-700;
 }
-
-.edit-site-new-template-modal {
-	.components-menu-item__button {
-		height: auto;
-		text-align: left;
-	}
-}
